Move modifiers into CalendarDay
