+++ css/styles.scss
@@ -1,3 +1,4 @@
+@import 'CalendarDay';
 @import 'CalendarMonth';
 @import 'CalendarMonthGrid';
 @import 'DayPicker';
