+++ test/components/CalendarDay_spec.jsx
@@ -4,7 +4,7 @@ import sinon from 'sinon-sandbox';
 import { shallow } from 'enzyme';
 import moment from 'moment';
 
-import CalendarDay from '../../src/components/CalendarDay';
+import CalendarDay, { getModifiersForDay } from '../../src/components/CalendarDay';
 
 describe('CalendarDay', () => {
   describe('#render', () => {
@@ -13,9 +13,9 @@ describe('CalendarDay', () => {
       expect(wrapper.is('.CalendarDay')).to.equal(true);
     });
 
-    it('has .CalendarDay__day class', () => {
+    it('has .CalendarDay class', () => {
       const wrapper = shallow(<CalendarDay />);
-      expect(wrapper.find('.CalendarDay__day')).to.have.lengthOf(1);
+      expect(wrapper.find('.CalendarDay')).to.have.lengthOf(1);
     });
 
     it('contains formatted day for single digit days', () => {
@@ -100,6 +100,35 @@ describe('CalendarDay', () => {
       const wrapper = shallow(<CalendarDay onDayMouseLeave={onMouseLeaveStub} />);
       wrapper.instance().onDayMouseLeave();
       expect(onMouseLeaveStub).to.have.property('callCount', 1);
+    });
+  });
+
+  describe('#getModifiersForDay', () => {
+    it('returns empty array if day is not passed in', () => {
+      const modifierKey = 'foo';
+      const modifiers = {};
+      modifiers[modifierKey] = () => true;
+
+      const filteredModifiers = getModifiersForDay(modifiers);
+      expect(filteredModifiers).to.have.lengthOf(0);
+    });
+
+    it('returns key for true modifier', () => {
+      const modifierKey = 'foo';
+      const modifiers = {};
+      modifiers[modifierKey] = () => true;
+
+      const filteredModifiers = getModifiersForDay(modifiers, moment());
+      expect(filteredModifiers).to.include(modifierKey);
+    });
+
+    it('does not return key for false modifier', () => {
+      const modifierKey = 'foo';
+      const modifiers = {};
+      modifiers[modifierKey] = () => false;
+
+      const filteredModifiers = getModifiersForDay(modifiers, moment());
+      expect(filteredModifiers).not.to.include(modifierKey);
     });
   });
 });
