+++ test/components/CalendarMonth_spec.jsx
@@ -4,7 +4,7 @@ import { shallow } from 'enzyme';
 import moment from 'moment';
 
 import { HORIZONTAL_ORIENTATION, VERTICAL_ORIENTATION } from '../../constants';
-import CalendarMonth, { getModifiersForDay } from '../../src/components/CalendarMonth';
+import CalendarMonth from '../../src/components/CalendarMonth';
 
 describe('CalendarMonth', () => {
   describe('#render', () => {
@@ -62,35 +62,6 @@ describe('CalendarMonth', () => {
         const wrapper = shallow(<CalendarMonth />);
         expect(wrapper.find('.js-CalendarMonth__grid')).to.have.lengthOf(1);
       });
-    });
-  });
-
-  describe('#getModifiersForDay', () => {
-    it('returns empty array if day is not passed in', () => {
-      const modifierKey = 'foo';
-      const modifiers = {};
-      modifiers[modifierKey] = () => true;
-
-      const filteredModifiers = getModifiersForDay(modifiers);
-      expect(filteredModifiers).to.have.lengthOf(0);
-    });
-
-    it('returns key for true modifier', () => {
-      const modifierKey = 'foo';
-      const modifiers = {};
-      modifiers[modifierKey] = () => true;
-
-      const filteredModifiers = getModifiersForDay(modifiers, moment());
-      expect(filteredModifiers).to.include(modifierKey);
-    });
-
-    it('does not return key for false modifier', () => {
-      const modifierKey = 'foo';
-      const modifiers = {};
-      modifiers[modifierKey] = () => false;
-
-      const filteredModifiers = getModifiersForDay(modifiers, moment());
-      expect(filteredModifiers).not.to.include(modifierKey);
     });
   });
 });
